GET functionality implemented for /movies/:id
The frontend can now access all movies and an individual movie identified by an ID.
Displaying this data still needs to be implemented and the search function, i.e. enter a string and the DB will be searched for titles that contain that string then this data will be displayed.